demo/advanced/demo2_security/setup/requirements.txt
flask
requests
pyjwt
bcrypt
sqlalchemy
psycopg2-binary
